+++ CHANGES.md
@@ -10,6 +10,7 @@
 - Update: Core Lightning (CLN - formerly C-lightning) v0.11.0.1 [details](https://github.com/ElementsProject/lightning/releases/tag/v0.11.0.1)
 - Update: LND v0.14.3 [details](https://github.com/lightningnetwork/lnd/releases/tag/v0.14.3-beta)
 - Update: RTL v0.12.3 [details](https://github.com/Ride-The-Lightning/RTL/releases/tag/v0.12.3)
+- Update: LNbits 0.8.0 [details](https://github.com/lnbits/lnbits-legend/releases/tag/0.8.0)
 - Update: C-lightningREST v0.7.0 [details](https://github.com/Ride-The-Lightning/c-lightning-REST/releases/tag/v0.7.0)
 - Update: CLBOSS 0.13A [details](https://github.com/ZmnSCPxj/clboss/releases/tag/0.13A)
 - Update: Channel Tools (chantools) v0.10.4 [details](https://github.com/guggero/chantools/blob/master/README.md)